Doing some debugging

+++ pages/index.vue
@@ -91,15 +91,28 @@ export default {
       this.todoInput = ''
     },
     addDoneTodos(index) {
-      this.doneTodos.push(this.todos.splice(index, 1))
+      const item = this.todos.splice(index, 1)
+      item.forEach((i) => {
+        this.doneTodos.push(i)
+      })
+      localStorage.setItem('todos', JSON.stringify(this.todos))
       localStorage.setItem('doneTodos', JSON.stringify(this.doneTodos))
     },
     editTodo(index) {
       this.todoInput = this.todos.splice(index, 1)
     },
-    deleteTodo(index) {
-      this.todos.splice(index, 1)
-      localStorage.setItem('todos', JSON.stringify(this.todos))
+    async deleteTodo(index) {
+      const confirmation = await prompt(
+        'Are you sure you want to delete this todo? Warning: this action is irreverible! Write Yes or No',
+        'Yes'
+      )
+      if (confirmation.toLowerCase === 'yes') {
+        this.todos.splice(index, 1)
+        localStorage.setItem('todos', JSON.stringify(this.todos))
+      } else if (confirmation.toLowerCase === 'no') {
+      } else {
+        alert('Value entered is invalid.')
+      }
     },
   },
   head() {
